fix: refactor events_file into Arg
This allows invoke and events commands to share the implementation and fixes #544
