+++ cmd/soroban-cli/src/commands/events.rs
@@ -1,16 +1,10 @@
-use chrono::{DateTime, NaiveDateTime, Utc};
-use clap::{arg, Parser};
-use itertools::Itertools;
-use std::{fs, io, path};
-use termcolor::{Color, ColorChoice, StandardStream, WriteColor};
-use termcolor_output::colored;
+use clap::{arg, command, Parser};
+use std::io;
 
-use soroban_env_host::{
-    events,
-    xdr::{self, ReadXdr, WriteXdr},
-};
+use soroban_env_host::xdr::{self, ReadXdr};
 
-use crate::commands::{HEADING_RPC, HEADING_SANDBOX};
+use super::config::{events_file, locator, network};
+use crate::commands::config::network::Network;
 use crate::{rpc, toid, utils};
 
 #[derive(Parser, Debug, Clone)]
@@ -39,26 +33,6 @@ pub struct Cmd {
     /// RPC).
     #[arg(short, long, default_value = "10")]
     count: usize,
-
-    /// RPC server endpoint
-    #[arg(long = "rpc-url",
-        env = "SOROBAN_RPC_URL",
-        help_heading = HEADING_RPC,
-        conflicts_with = "events_file",
-        required_unless_present="events_file",
-    )]
-    rpc_url: Option<String>,
-
-    /// Local event store (likely generated by `invoke`) to pull events from
-    #[arg(
-        long,
-        value_name = "PATH",
-        env = "SOROBAN_EVENTS_FILE",
-        help_heading = HEADING_SANDBOX,
-        conflicts_with = "rpc_url",
-        required_unless_present="rpc_url",
-    )]
-    events_file: Option<std::path::PathBuf>,
 
     /// A set of (up to 5) contract IDs to filter events on. This parameter can
     /// be passed multiple times, e.g. `--id abc --id def`, or passed with
@@ -90,7 +64,7 @@ pub struct Cmd {
     /// into a single filter (i.e. combination of type, IDs, and topics).
     #[arg(
         long = "topic",
-        num_args = 1..=5,
+        num_args = 1..=4,
         help_heading = "FILTERS"
     )]
     topic_filters: Vec<String>,
@@ -103,6 +77,15 @@ pub struct Cmd {
         help_heading = "FILTERS"
     )]
     event_type: rpc::EventType,
+
+    #[command(flatten)]
+    locator: locator::Args,
+
+    #[command(flatten)]
+    network: network::Args,
+
+    #[command(flatten)]
+    events_file: events_file::Args,
 }
 
 #[derive(thiserror::Error, Debug)]
@@ -160,6 +143,15 @@ pub enum Error {
 
     #[error(transparent)]
     Serde(#[from] serde_json::Error),
+
+    #[error(transparent)]
+    Network(#[from] network::Error),
+
+    #[error(transparent)]
+    EventsFile(#[from] events_file::Error),
+
+    #[error(transparent)]
+    Locator(#[from] locator::Error),
 }
 
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, clap::ValueEnum)]
@@ -202,11 +194,11 @@ impl Cmd {
             }
         }
 
-        let response = match (self.rpc_url.as_ref(), self.events_file.as_ref()) {
-            (Some(rpc_url), _) => self.run_against_rpc_server(rpc_url, start).await,
-            (_, Some(path)) => self.run_in_sandbox(path, start),
-            // should never happen because of required_unless_present flags
-            _ => return Err(Error::MissingTarget),
+        let response = if self.network.is_no_network() {
+            let Network { rpc_url, .. } = self.network.get(&self.locator)?;
+            self.run_against_rpc_server(&rpc_url, start).await
+        } else {
+            self.run_in_sandbox(start)
         }?;
 
         for event in &response.events {
@@ -225,8 +217,8 @@ impl Cmd {
                         })?,
                     );
                 }
-                OutputFormat::Plain => print_event(event)?,
-                OutputFormat::Pretty => pretty_print_event(event)?,
+                OutputFormat::Plain => println!("{event}"),
+                OutputFormat::Pretty => event.pretty_print()?,
             }
         }
         println!("Latest Ledger: {}", response.latest_ledger);
@@ -262,17 +254,7 @@ impl Cmd {
             .map_err(Error::Rpc)
     }
 
-    pub fn run_in_sandbox(
-        &self,
-        path: &path::PathBuf,
-        start: rpc::EventStart,
-    ) -> Result<rpc::GetEventsResponse, Error> {
-        if !path.exists() {
-            return Err(Error::InvalidFile {
-                path: path.to_str().unwrap().to_string(),
-            });
-        }
-
+    pub fn run_in_sandbox(&self, start: rpc::EventStart) -> Result<rpc::GetEventsResponse, Error> {
         let count: usize = if self.count == 0 {
             std::usize::MAX
         } else {
@@ -281,468 +263,42 @@ impl Cmd {
 
         let start_cursor = match start {
             rpc::EventStart::Ledger(l) => (toid::Toid::new(l, 0, 0).into(), -1),
-            rpc::EventStart::Cursor(c) => parse_cursor(&c)?,
+            rpc::EventStart::Cursor(c) => rpc::parse_cursor(&c)?,
         };
-
-        let file = read(path).map_err(|err| Error::CannotReadFile {
-            path: path.to_str().unwrap().to_string(),
-            error: err.to_string(),
-        })?;
+        let path = self.locator.config_dir()?;
+        let file = self.events_file.read(&path)?;
 
         // Read the JSON events from disk and find the ones that match the
         // contract ID filter(s) that were passed in.
         Ok(rpc::GetEventsResponse {
-            events: file
-                .events
-                .iter()
-                .filter(|evt| match parse_cursor(&evt.id) {
-                    Ok(event_cursor) => event_cursor > start_cursor,
-                    Err(e) => {
-                        eprintln!("error parsing key 'ledger': {e:?}");
-                        eprintln!(
-                            "your sandbox events file ('{}') may be corrupt, consider deleting it",
-                            path.to_str().unwrap(),
-                        );
-                        eprintln!("ignoring this event: {evt:#?}");
-
-                        false
-                    }
-                })
-                .filter(|evt| {
-                    // Contract ID filter(s) are optional, so we should render all
-                    // events if they're omitted.
-                    self.contract_ids.is_empty()
-                        || self.contract_ids.iter().any(|id| *id == evt.contract_id)
-                })
-                .filter(|evt| {
-                    // Like before, no topic filters means pass everything through.
-                    self.topic_filters.is_empty() ||
-                // Reminder: All of the topic filters are part of a single
-                // filter object, and each one contains segments, so we need to
-                // apply all of them to the given event.
-                self.topic_filters
-                    .iter()
-                    // quadratic, but both are <= 5 long
-                    .any(|f| {
-                        does_topic_match(
-                            &evt.topic,
-                            // misc. Rust nonsense: make a copy over the given
-                            // split filter, because passing a slice of
-                            // references is too much for this language to
-                            // handle
-                            &f.split(',')
-                            .map(std::string::ToString::to_string)
-                            .collect::<Vec<String>>()
-                        )
-                    })
-                })
-                .take(count)
-                .cloned()
-                .collect::<Vec<rpc::Event>>(),
+            events: file.filter_events(
+                &path,
+                start_cursor,
+                &self.contract_ids,
+                &self.topic_filters,
+                count,
+            )?,
             latest_ledger: file.latest_ledger,
         })
     }
 }
 
-fn parse_cursor(c: &str) -> Result<(u64, i32), Error> {
-    let (toid_part, event_index) = c.split('-').collect_tuple().ok_or(Error::InvalidCursor)?;
-    let toid_part: u64 = toid_part.parse().map_err(|_| Error::InvalidCursor)?;
-    let start_index: i32 = event_index.parse().map_err(|_| Error::InvalidCursor)?;
-    Ok((toid_part, start_index))
-}
-
-// Determines whether or not a particular filter matches a topic based on the
-// same semantics as the RPC server:
-//
-//  - for an exact segment match, the filter is a base64-encoded ScVal
-//  - for a wildcard, single-segment match, the string "*" matches exactly one
-//    segment
-//
-// The expectation is that a `filter` is a comma-separated list of segments that
-// has previously been validated, and `topic` is the list of segments applicable
-// for this event.
-//
-// [API
-// Reference](https://docs.google.com/document/d/1TZUDgo_3zPz7TiPMMHVW_mtogjLyPL0plvzGMsxSz6A/edit#bookmark=id.35t97rnag3tx)
-// [Code
-// Reference](https://github.com/stellar/soroban-tools/blob/bac1be79e8c2590c9c35ad8a0168aab0ae2b4171/cmd/soroban-rpc/internal/methods/get_events.go#L182-L203)
-pub fn does_topic_match(topic: &[String], filter: &[String]) -> bool {
-    filter.len() == topic.len()
-        && filter
-            .iter()
-            .enumerate()
-            .all(|(i, s)| *s == "*" || topic[i] == *s)
-}
-
-pub fn print_event(event: &rpc::Event) -> Result<(), Box<dyn std::error::Error>> {
-    println!(
-        "Event {} [{}]:",
-        event.paging_token,
-        event.event_type.to_ascii_uppercase()
-    );
-    println!(
-        "  Ledger:   {} (closed at {})",
-        event.ledger, event.ledger_closed_at
-    );
-    println!("  Contract: {}", event.contract_id);
-    println!("  Topics:");
-    for topic in &event.topic {
-        let scval = xdr::ScVal::from_xdr_base64(topic)?;
-        println!("            {scval:?}");
-    }
-    let scval = xdr::ScVal::from_xdr_base64(&event.value.xdr)?;
-    println!("  Value:    {scval:?}");
-
-    Ok(())
-}
-
-pub fn pretty_print_event(event: &rpc::Event) -> Result<(), Box<dyn std::error::Error>> {
-    let mut stdout = StandardStream::stdout(ColorChoice::Auto);
-    if !stdout.supports_color() {
-        print_event(event)?;
-        return Ok(());
-    }
-
-    let color = match event.event_type.as_str() {
-        "system" => Color::Yellow,
-        _ => Color::Blue,
-    };
-    colored!(
-        stdout,
-        "{}Event{} {}{}{} [{}{}{}{}]:\n",
-        bold!(true),
-        bold!(false),
-        fg!(Some(Color::Green)),
-        event.paging_token,
-        reset!(),
-        bold!(true),
-        fg!(Some(color)),
-        event.event_type.to_ascii_uppercase(),
-        reset!(),
-    )?;
-
-    colored!(
-        stdout,
-        "  Ledger:   {}{}{} (closed at {}{}{})\n",
-        fg!(Some(Color::Green)),
-        event.ledger,
-        reset!(),
-        fg!(Some(Color::Green)),
-        event.ledger_closed_at,
-        reset!(),
-    )?;
-
-    colored!(
-        stdout,
-        "  Contract: {}0x{}{}\n",
-        fg!(Some(Color::Green)),
-        event.contract_id,
-        reset!(),
-    )?;
-
-    colored!(stdout, "  Topics:\n")?;
-    for topic in &event.topic {
-        let scval = xdr::ScVal::from_xdr_base64(topic)?;
-        colored!(
-            stdout,
-            "            {}{:?}{}\n",
-            fg!(Some(Color::Green)),
-            scval,
-            reset!(),
-        )?;
-    }
-
-    let scval = xdr::ScVal::from_xdr_base64(&event.value.xdr)?;
-    colored!(
-        stdout,
-        "  Value: {}{:?}{}\n",
-        fg!(Some(Color::Green)),
-        scval,
-        reset!(),
-    )?;
-
-    Ok(())
-}
-
-/// Returns a list of events from the on-disk event store, which stores events
-/// exactly as they'd be returned by an RPC server.
-pub fn read(path: &std::path::PathBuf) -> Result<rpc::GetEventsResponse, Error> {
-    let reader = std::fs::OpenOptions::new().read(true).open(path)?;
-    Ok(serde_json::from_reader(reader)?)
-}
-
-/// Reads the existing event file, appends the new events, and writes it all to
-/// disk. Note that this almost certainly isn't safe to call in parallel.
-pub fn commit(
-    new_events: &[events::HostEvent],
-    ledger_info: &soroban_ledger_snapshot::LedgerSnapshot,
-    output_file: &std::path::PathBuf,
-) -> Result<(), Error> {
-    // Create the directory tree if necessary, since these are unlikely to be
-    // the first events.
-    if let Some(dir) = output_file.parent() {
-        if !dir.exists() {
-            fs::create_dir_all(dir)?;
-        }
-    }
-
-    let mut events: Vec<rpc::Event> = if path::Path::exists(output_file) {
-        let mut file = fs::OpenOptions::new().read(true).open(output_file)?;
-        let payload: rpc::GetEventsResponse = serde_json::from_reader(&mut file)?;
-        payload.events
-    } else {
-        vec![]
-    };
-
-    for (i, event) in new_events.iter().enumerate() {
-        let contract_event = match &event.event {
-            events::Event::Contract(e) => e,
-            events::Event::Debug(e) => {
-                return Err(Error::Generic(
-                    format!("debug events unsupported: {e:#?}").into(),
-                ))
-            }
-            events::Event::StructuredDebug(e) => {
-                return Err(Error::Generic(
-                    format!("structured debug events unsupported: {e:#?}").into(),
-                ))
-            }
-        };
-
-        let topics = match &contract_event.body {
-            xdr::ContractEventBody::V0(e) => &e.topics,
-        }
-        .iter()
-        .map(xdr::WriteXdr::to_xdr_base64)
-        .collect::<Result<Vec<String>, _>>()?;
-
-        // stolen from
-        // https://github.com/stellar/soroban-tools/blob/main/cmd/soroban-rpc/internal/methods/get_events.go#L264
-        let id = format!(
-            "{}-{:010}",
-            toid::Toid::new(
-                ledger_info.sequence_number,
-                // we should technically inject the tx order here from the
-                // ledger info, but the sandbox does one tx/op per ledger
-                // anyway, so this is a safe assumption
-                1,
-                1,
-            )
-            .to_paging_token(),
-            i + 1
-        );
-
-        // Misc. timestamp to RFC 3339-formatted datetime nonsense, with an
-        // absurd amount of verbosity because every edge case needs its own
-        // chain of error-handling methods.
-        //
-        // Reference: https://stackoverflow.com/a/50072164
-        let ts: i64 = ledger_info
-            .timestamp
-            .try_into()
-            .map_err(|_e| Error::InvalidTimestamp {
-                ts: ledger_info.timestamp.to_string(),
-            })?;
-        let ndt =
-            NaiveDateTime::from_timestamp_opt(ts, 0).ok_or_else(|| Error::InvalidTimestamp {
-                ts: ledger_info.timestamp.to_string(),
-            })?;
-
-        let dt: DateTime<Utc> = DateTime::from_utc(ndt, Utc);
-
-        let cereal_event = rpc::Event {
-            event_type: match contract_event.type_ {
-                xdr::ContractEventType::Contract => "contract",
-                xdr::ContractEventType::System => "system",
-                xdr::ContractEventType::Diagnostic => "diagnostic",
-            }
-            .to_string(),
-            paging_token: id.clone(),
-            id,
-            ledger: ledger_info.sequence_number.to_string(),
-            ledger_closed_at: dt.format("%Y-%m-%dT%H:%M:%SZ").to_string(),
-            contract_id: hex::encode(
-                contract_event
-                    .contract_id
-                    .as_ref()
-                    .unwrap_or(&xdr::Hash([0; 32])),
-            ),
-            topic: topics,
-            value: rpc::EventValue {
-                xdr: match &contract_event.body {
-                    xdr::ContractEventBody::V0(e) => &e.data,
-                }
-                .to_xdr_base64()?,
-            },
-        };
-
-        events.push(cereal_event);
-    }
-
-    let mut file = std::fs::OpenOptions::new()
-        .create(true)
-        .write(true)
-        .truncate(true)
-        .open(output_file)?;
-
-    serde_json::to_writer_pretty(
-        &mut file,
-        &rpc::GetEventsResponse {
-            events,
-            latest_ledger: ledger_info.sequence_number,
-        },
-    )?;
-
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
+    use std::path;
+
     use assert_fs::NamedTempFile;
+    use soroban_env_host::events;
 
     use super::*;
 
-    #[test]
-    // Taken from [RPC server
-    // tests](https://github.com/stellar/soroban-tools/blob/main/cmd/soroban-rpc/internal/methods/get_events_test.go#L21).
-    fn test_does_topic_match() {
-        struct TestCase<'a> {
-            name: &'a str,
-            filter: Vec<&'a str>,
-            includes: Vec<Vec<&'a str>>,
-            excludes: Vec<Vec<&'a str>>,
-        }
-
-        let xfer = "AAAABQAAAAh0cmFuc2Zlcg==";
-        let number = "AAAAAQB6Mcc=";
-        let star = "*";
-
-        for tc in vec![
-            // No filter means match nothing.
-            TestCase {
-                name: "<empty>",
-                filter: vec![],
-                includes: vec![],
-                excludes: vec![vec![xfer]],
-            },
-            // "*" should match "transfer/" but not "transfer/transfer" or
-            // "transfer/amount", because * is specified as a SINGLE segment
-            // wildcard.
-            TestCase {
-                name: "*",
-                filter: vec![star],
-                includes: vec![vec![xfer]],
-                excludes: vec![vec![xfer, xfer], vec![xfer, number]],
-            },
-            // "*/transfer" should match anything preceding "transfer", but
-            // nothing that isn't exactly two segments long.
-            TestCase {
-                name: "*/transfer",
-                filter: vec![star, xfer],
-                includes: vec![vec![number, xfer], vec![xfer, xfer]],
-                excludes: vec![
-                    vec![number],
-                    vec![number, number],
-                    vec![number, xfer, number],
-                    vec![xfer],
-                    vec![xfer, number],
-                    vec![xfer, xfer, xfer],
-                ],
-            },
-            // The inverse case of before: "transfer/*" should match any single
-            // segment after a segment that is exactly "transfer", but no
-            // additional segments.
-            TestCase {
-                name: "transfer/*",
-                filter: vec![xfer, star],
-                includes: vec![vec![xfer, number], vec![xfer, xfer]],
-                excludes: vec![
-                    vec![number],
-                    vec![number, number],
-                    vec![number, xfer, number],
-                    vec![xfer],
-                    vec![number, xfer],
-                    vec![xfer, xfer, xfer],
-                ],
-            },
-            // Here, we extend to exactly two wild segments after transfer.
-            TestCase {
-                name: "transfer/*/*",
-                filter: vec![xfer, star, star],
-                includes: vec![vec![xfer, number, number], vec![xfer, xfer, xfer]],
-                excludes: vec![
-                    vec![number],
-                    vec![number, number],
-                    vec![number, xfer],
-                    vec![number, xfer, number, number],
-                    vec![xfer],
-                    vec![xfer, xfer, xfer, xfer],
-                ],
-            },
-            // Here, we ensure wildcards can be in the middle of a filter: only
-            // exact matches happen on the ends, while the middle can be
-            // anything.
-            TestCase {
-                name: "transfer/*/number",
-                filter: vec![xfer, star, number],
-                includes: vec![vec![xfer, number, number], vec![xfer, xfer, number]],
-                excludes: vec![
-                    vec![number],
-                    vec![number, number],
-                    vec![number, number, number],
-                    vec![number, xfer, number],
-                    vec![xfer],
-                    vec![number, xfer],
-                    vec![xfer, xfer, xfer],
-                    vec![xfer, number, xfer],
-                ],
-            },
-        ] {
-            for topic in tc.includes {
-                assert!(
-                    does_topic_match(
-                        &topic
-                            .iter()
-                            .map(std::string::ToString::to_string)
-                            .collect::<Vec<String>>(),
-                        &tc.filter
-                            .iter()
-                            .map(std::string::ToString::to_string)
-                            .collect::<Vec<String>>()
-                    ),
-                    "test: {}, topic ({:?}) should be matched by filter ({:?})",
-                    tc.name,
-                    topic,
-                    tc.filter
-                );
-            }
-
-            for topic in tc.excludes {
-                assert!(
-                    !does_topic_match(
-                        // make deep copies of the vecs
-                        &topic
-                            .iter()
-                            .map(std::string::ToString::to_string)
-                            .collect::<Vec<String>>(),
-                        &tc.filter
-                            .iter()
-                            .map(std::string::ToString::to_string)
-                            .collect::<Vec<String>>()
-                    ),
-                    "test: {}, topic ({:?}) should NOT be matched by filter ({:?})",
-                    tc.name,
-                    topic,
-                    tc.filter
-                );
-            }
-        }
-    }
-
+    use events_file::Args;
     #[test]
     fn test_does_event_serialization_match() {
         let temp = NamedTempFile::new("events.json").unwrap();
-
+        let events_file = Args {
+            events_file: Some(temp.to_path_buf()),
+        };
         // Make a couple of fake events with slightly different properties and
         // write them to disk, then read the serialized versions from disk and
         // ensure the properties match.
@@ -774,7 +330,7 @@ mod tests {
             },
         ];
 
-        let snapshot = soroban_ledger_snapshot::LedgerSnapshot {
+        let ledger_info = soroban_ledger_snapshot::LedgerSnapshot {
             protocol_version: 1,
             sequence_number: 2, // this is the only value that matters
             timestamp: 3,
@@ -783,9 +339,9 @@ mod tests {
             ledger_entries: vec![],
         };
 
-        commit(&events, &snapshot, &temp.to_path_buf()).unwrap();
+        events_file.commit(&events, &ledger_info, &temp).unwrap();
 
-        let file = read(&temp.to_path_buf()).unwrap();
+        let file = events_file.read(&std::env::current_dir().unwrap()).unwrap();
         assert_eq!(file.events.len(), 2);
         assert_eq!(file.events[0].ledger, "2");
         assert_eq!(file.events[1].ledger, "2");
@@ -800,8 +356,10 @@ mod tests {
         // correct event format (for the purposes of human readability).
         let filename =
             path::PathBuf::from("../crates/soroban-test/tests/fixtures/test-jsons/get-events.json");
-
-        let result = read(&filename);
+        let events_file = Args {
+            events_file: Some(filename),
+        };
+        let result = events_file.read(&std::env::current_dir().unwrap());
         println!("{result:?}");
         assert!(result.is_ok());
     }
